Rename folder with set and get functions
