+++ src/components/Database/AddNewJob.js
@@ -2,9 +2,6 @@ import React, { useEffect, useState } from "react";
 import "./AddNewJob.css";
 import useSettings from "../../core/customHooks/useSettings";
 import combineAllJobInputsInOneVariable from "../../core/combineAllJobInputsInOneVariable";
-import getListOfSourcesFromFB from "../../core/getFromFBFuctions/getListOfSourcesFromFB";
-import getListOfCitiesFromFB from "../../core/getFromFBFuctions/getListOfCitiesFromFB";
-import getListOfResultsFromFB from "../../core/getFromFBFuctions/getListOfResultsFromFB";
 
 const AddNewJob = ({ handleAddJobToListSubmit, user }) => {
   const [companyName, setCompanyName] = useState("");
