+++ src/core/getFromFBFunctions/fetchListOfCompanyInformantionUserApplied.js
