+++ src/core/getFromFBFunctions/fetchListOfCompanyNamesUserApplied.js
