+++ src/core/getFromFBFunctions/getListOfCitiesFromFB.js
