+++ src/core/getFromFBFunctions/getListOfResultsFromFB.js
