+++ src/core/getFromFBFunctions/getListOfSourcesFromFB.js
