+++ src/core/setToFBFunctions/setCompanyNamesListToFB.js
