+++ src/core/setToFBFunctions/setDefaultNewUserSettingsToFB.js
