+++ src/core/setToFBFunctions/setJobsInformationListToFB.js
