+++ src/core/setToFBFunctions/setUpdatedListOfCitiesToFB.js
