+++ src/core/setToFBFunctions/setUpdatedListOfResultsToFB.js
