+++ src/core/setToFBFunctions/setUpdatedListOfSourcesToFB.js
